Refactor : déplacement du service d'autentification dans common
