+++ src/app/app.module.ts
@@ -25,7 +25,6 @@ import {MatCheckboxModule} from "@angular/material/checkbox";
 import {MatSnackBarModule} from '@angular/material/snack-bar';
 import {ApiModule, Configuration} from "./apis/litopia-api";
 import {environment} from '../environments/environment';
-import {AuthPopupComponent} from './auth/auth-popup/auth-popup.component';
 import {MatDialogModule} from "@angular/material/dialog";
 import {MembresComponent} from './pages/membres/membres.component';
 import {AccueilModule} from "./accueil/accueil.module";
@@ -40,7 +39,6 @@ import {NousRejoindreModule} from "./nous-rejoindre/nous-rejoindre.module";
     NotFoundComponent,
     FooterComponent,
     ReglementComponent,
-    AuthPopupComponent,
     MembresComponent
   ],
   imports: [
