+++ src/app/common/components/auth-popup/auth-popup.component.ts
