+++ src/app/common/litopia-common.module.ts
@@ -5,6 +5,11 @@ import {
 } from "./components/picture-content-delivery/picture-content-delivery.component";
 import {AppearDirective} from "./directives/appear.directive";
 import {PageHeaderComponent} from "./components/page-header/page-header.component";
+import {AuthPopupComponent} from "./components/auth-popup/auth-popup.component";
+import {MatCheckboxModule} from "@angular/material/checkbox";
+import {FormsModule} from "@angular/forms";
+import {MatDialogModule} from "@angular/material/dialog";
+import {MatButtonModule} from "@angular/material/button";
 
 
 
@@ -13,9 +18,14 @@ import {PageHeaderComponent} from "./components/page-header/page-header.componen
     PictureContentDeliveryComponent,
     PageHeaderComponent,
     AppearDirective,
+    AuthPopupComponent,
   ],
   imports: [
     CommonModule,
+    MatCheckboxModule,
+    FormsModule,
+    MatDialogModule,
+    MatButtonModule,
   ],
   exports: [
     AppearDirective,
